Add workaround for handling page titles with slashes
On Wiki Education Dashboard, we noticed that authorship data would not load for the Wikipedia article "Post-9/11". This is because Django is matching the url-pattern for a page_title followed by a rev_id, rather than the correct title without a rev_id. (When a rev_id is also provided, it was working properly.)

This allows for the using rev_id 0 as an equivalent to a request without a rev_id, so that url matching in whocolor/urls.py can work for "Post-9/11" and similarly titled articles when called for the (default) latest revision rather than a specified revision.