Add itwiki support
This intentionally doesn't add itwiki to event_streams.py yet, because
it probably makes sense to let the initial import finish before enabling
realtime processing. In production, this would create pickle files on
the fly, on top of the heavy operation of importing the XML dumps,
leading to redundancy and possibly a collision if the pickle file
already exists by the time the import script gets to that article.

(Completely untested theory, just playing it safe :)

Bug: T270490