Delete pickle file after article has been deleted
This also fixes a bug where the language code is assumed to be the
first two characters of the domain name. Change this to go by the
server_name from EventStreams (which is the domain), and use the
subdomain as the language code.

The db list for event streams has been moved to the EVENT_STREAMS_WIKIS
setting, and the language lists have been alphabetized.

Bug: T333390